Pad 3 column layout on the inside
Previously, I would pad the 3 column layout around all of the windows.
Then I moved polybar to the top of the screen, which looks much better.
After this move, The interface looks nicer without gaps on the top,
bottom and sides of the tiled layout. This has a side effect of making
both tiled and full-screen layouts the same when 1 window is present.