Allow empty CRTC map
Turns out that when you disable a monitor at the coordinate 0,0, the X
window system will send something such as the following:

    ID  Crtc ID  X  Y  Width Height
    0   63       0  0  0     0
    1   64       0  0  0     0
    2   63       0  0  1920  1280

This would cause the prior version of lanta to crash at transation # 1
saying that it must have a screen. This change allows all three of these
transations to go through in the hopes that the X server will enable a
monitor in the future.